# perf: fix SIMD-inlining